+++ Octokit.Reactive/Clients/ObservableActionsArtifactsClient.cs
@@ -27,25 +27,25 @@ namespace Octokit.Reactive
         }
 
         /// <inheritdoc/>
-        public IObservable<Artifact> GetArtifact(string owner, string repository, int artifactId)
+        public IObservable<Artifact> GetArtifact(string owner, string repository, long artifactId)
         {
             return _client.GetArtifact(owner, repository, artifactId).ToObservable();
         }
 
         /// <inheritdoc/>
-        public IObservable<Unit> DeleteArtifact(string owner, string repository, int artifactId)
+        public IObservable<Unit> DeleteArtifact(string owner, string repository, long artifactId)
         {
             return _client.DeleteArtifact(owner, repository, artifactId).ToObservable();
         }
 
         /// <inheritdoc/>
-        public IObservable<Stream> DownloadArtifact(string owner, string repository, int artifactId, string archiveFormat)
+        public IObservable<Stream> DownloadArtifact(string owner, string repository, long artifactId, string archiveFormat)
         {
             return _client.DownloadArtifact(owner, repository, artifactId, archiveFormat).ToObservable();
         }
         
         /// <inheritdoc/>
-        public IObservable<ListArtifactsResponse> ListWorkflowArtifacts(string owner, string repository, int runId, ListArtifactsRequest listArtifactsRequest = null)
+        public IObservable<ListArtifactsResponse> ListWorkflowArtifacts(string owner, string repository, long runId, ListArtifactsRequest listArtifactsRequest = null)
         {
             return _client.ListWorkflowArtifacts(owner, repository, runId, listArtifactsRequest).ToObservable();
         }
